added all sections
added sections to drawer
adjusted styling for each drawer component
dynamically rendering present content
