+++ src/components/CoachProfileDrawer/components/contentDivider.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Divider } from '@chakra-ui/react';
+
+const ContentDivider = () => {
+  return (
+    <Divider borderColor='brand.300' w='90%' />
+  )
+}
+
+export default ContentDivider;
