+++ src/components/TagIcon/index.tsx
@@ -1,15 +1,7 @@
 import React from "react";
 import { Icon } from "@chakra-ui/react";
-import {
-  AnchorIcon,
-  BookIcon,
-  CrownIcon,
-  DroneIcon,
-  FireworksIcon,
-  ShieldIcon,
-  SubsetIcon,
-} from "../../assets/Icons/StyleIcons";
 import { IconProps } from "@chakra-ui/react";
+import { iconComponents } from "./utils";
 
 type TagIconProps = IconProps & {
   isHovered?: boolean;
@@ -19,24 +11,20 @@ type TagIconProps = IconProps & {
   icon: string;
 };
 
-const iconComponents = {
-  shield: ShieldIcon,
-  book: BookIcon,
-  drone: DroneIcon,
-  anchor: AnchorIcon,
-  subset: SubsetIcon,
-  fireworks: FireworksIcon,
-  crown: CrownIcon,
-};
-
-export const TagIcon = ({ isHovered, isSelected, isSelectable, icon, isBgWhite, ...rest }: TagIconProps) => {
-
+export const TagIcon = ({
+  isHovered,
+  isSelected,
+  isSelectable,
+  icon,
+  isBgWhite,
+  ...rest
+}: TagIconProps) => {
   const IconComponent = iconComponents[icon];
 
   return (
     <Icon
       as={IconComponent}
-      mr={2}
+      mr={1}
       isHovered={isHovered}
       isSelected={isSelected}
       isSelectable={isSelectable}
