+++ src/components/TagIcon/utils.ts
@@ -1,0 +1,19 @@
+import {
+  AnchorIcon,
+  BookIcon,
+  CrownIcon,
+  DroneIcon,
+  FireworksIcon,
+  ShieldIcon,
+  SubsetIcon,
+} from "../../assets/Icons/StyleIcons";
+
+export const iconComponents = {
+  shield: ShieldIcon,
+  book: BookIcon,
+  drone: DroneIcon,
+  anchor: AnchorIcon,
+  subset: SubsetIcon,
+  fireworks: FireworksIcon,
+  crown: CrownIcon,
+};
